home L

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: rbalbous <rbalbous@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/21 16:49:43 by rbalbous          #+#    #+#             */
-/*   Updated: 2018/01/06 18:28:42 by rbalbous         ###   ########.fr       */
+/*   Updated: 2018/01/07 12:56:43 by rbalbous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -174,16 +174,16 @@ int				pf_infinitec(long double d, t_flags *flags, t_var *var);
 int				pf_nanc(t_flags *flags, t_var *var);
 
 int				pf_e(t_flags *flags, t_var *var, va_list *ap);
-int				pf_spe_e(t_flags *flags, t_var *var, double d, int count);
+int				pf_spe_e(t_flags *flags, t_var *var, long double d, int count);
 int				pf_cap_e(t_flags *flags, t_var *var, va_list *ap);
-int				pf_spe_cap_e(t_flags *flags, t_var *var, double d, int count);
-int				pf_tosc(double *d);
+int				pf_spe_cap_e(t_flags *flags, t_var *var, long double d, int count);
+int				pf_tosc(long double *d);
 
 int				pf_a(t_flags *flags, t_var *var, va_list *ap);
-void			pf_ftoa_hexa(double n, t_flags *flags, t_var *var);
+void			pf_ftoa_hexa(long double n, t_flags *flags, t_var *var);
 void			pf_initoa(char modu[16]);
 int				pf_cap_a(t_flags *flags, t_var *var, va_list *ap);
-int				pf_toa(double *d);
+int				pf_toa(long double *d);
 
 int				pf_g(t_flags *flags, t_var *var, va_list *ap);
